int → ptrdiff_t for read results
* src/fileio.c (Finsert_file_contents): Use ptrdiff_t, not int, to
store results returned by emacs_fd_read, which return ptrdiff_t.
This doesn’t fix any bugs since the values happen to fit in int,
but is a bit clearer and may simplify future changes.

+++ src/fileio.c
@@ -4501,7 +4501,7 @@ by calling `format-decode', which see.  */)
 
       while (!giveup_match_end)
 	{
-	  int total_read, nread, bufpos, trial;
+	  ptrdiff_t total_read, nread, bufpos, trial;
 	  off_t curpos;
 
 	  /* At what file position are we now scanning?  */
